chore: move safe into /safe and split
